Implemented the changes requested in the PR
